dashboards/pyproject.toml
[project]
name = "energyintel-dashboards"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "streamlit>=1.52.2",
]
